feat(parsers): adding dynamic templates
A dynamic template can be declared by prefix an existing template with the keyword dynamic, for
example "{{tp.dynamic.file.last_modified_date()}}
The current solution is not perfect, the page
preview is getting cached in obsidian. This causes our dynamic template to only be triggered one
time at the start / reload of obsidian. Will need to improve that somehow.

re #63