Use user styles for syntax highlighting
This commit changes the styling used by Templater's syntax highlighting
to use user-defined colours under the Obsidian 1.0 style sheet. It aims
to match the built-in JavaScript syntax highlighting as closely as
possible.

It is backwards compatible in that colours will continue to work for
earlier versions of obsidian. However, some classes have been
consolidated to provide a more coherent style experience.